Update theme to work with Drupal 7
This update allows the research guide theme to work with Drupal 7.
It should be noted that only the parts that the research guides
system actually makes use of were the only parts that were updated.

It is probably advisable to create an entirely new theme, or at
least base the research guide theme off a more modern theme in the
future, as bluemarine is very out of date.