Include a patch for the customizations of noderefcreate module
The noderefcreate module changes the functionality that we used to
have in Drupal 6. If there is a node that has an identical name to
the one that you are trying to create, it will default to referring
to that node instead of creating a new one. This is not ideal since
many subject guide pages may have the same name. This patch will fix
this issue